kernel: qos_id: moved to common
Changed type to int and QOS_id_BAD to -1 to be consistend with
cep_id, port_id, etc. Moved to common.h/c for the same reasons.
