+++ linux/net/rina/pci.c
@@ -30,7 +30,6 @@
 #include "utils.h"
 #include "debug.h"
 #include "pdu.h"
-#include "qos.h"
 
 /* FIXME: These externs have to disappear from here */
 struct buffer * buffer_create_with_gfp(gfp_t  flags,
