+++ linux/net/rina/pci.h
@@ -27,7 +27,6 @@
 #include <linux/types.h>
 
 #include "common.h"
-#include "qos.h"
 #include "buffer.h"
 #include "pdu.h"
