--- linux/net/rina/qos.h
@@ -1,32 +1,0 @@
-/*
- * QoS
- *
- *    Francesco Salvestrini <f.salvestrini@nextworks.it>
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
- */
-
-#ifndef RINA_QOS_H
-#define RINA_QOS_H
-
-#include <linux/types.h>
-
-/* ALWAYS use this function to check if the id looks good */
-bool     is_qos_id_ok(qos_id_t id);
-
-/* ALWAYS use this function to get a bad id */
-qos_id_t qos_id_bad(void);
-
-#endif
